+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/player_skills/PlayerSkillsAdder.java
@@ -1,5 +1,6 @@
 package com.robertx22.age_of_exile.aoe_data.database.player_skills;
 
+import com.robertx22.age_of_exile.aoe_data.database.player_skills.skills.AlchemyAdder;
 import com.robertx22.age_of_exile.aoe_data.database.player_skills.skills.FarmingAdder;
 import com.robertx22.age_of_exile.aoe_data.database.player_skills.skills.MiningAdder;
 import com.robertx22.age_of_exile.database.registry.ISlashRegistryInit;
@@ -11,6 +12,7 @@ public class PlayerSkillsAdder implements ISlashRegistryInit {
 
         MiningAdder.createSkill();
         FarmingAdder.createSkill();
+        AlchemyAdder.createSkill();
 
     }
 }