+++ src/main/java/com/robertx22/age_of_exile/database/data/player_skills/PlayerSkill.java
@@ -30,6 +30,8 @@ public class PlayerSkill implements ISerializedRegistryEntry<PlayerSkill>, IAuto
 
     public List<BlockBreakExp> block_break_exp = new ArrayList<>();
 
+    public List<ItemCraftExp> item_craft_exp = new ArrayList<>();
+
     public Identifier getIcon() {
         return Ref.id("textures/gui/skills/icons/" + id + ".png");
     }