+++ src/main/java/com/robertx22/age_of_exile/gui/screens/player_skills/PlayerSkillsScreen.java
@@ -121,7 +121,7 @@ public class PlayerSkillsScreen extends BaseScreen implements INamedScreen {
             String xptext = exp + "/" + needed;
             String nametext = CLOC.translate(skill.word.locName());
 
-            TextUtils.renderText(matrix, 1.25, nametext, x + BUTTON_SIZE_X / 2, y - 15, Formatting.GOLD);
+            TextUtils.renderText(matrix, 1, nametext, x + BUTTON_SIZE_X / 2, y - 10, Formatting.GOLD);
             TextUtils.renderText(matrix, 0.8, lvltext, x + BUTTON_SIZE_X / 2, (int) (y + BUTTON_SIZE_Y * 1.2F), Formatting.YELLOW);
             TextUtils.renderText(matrix, 0.75, xptext, x + BUTTON_SIZE_X / 2, (int) (y + BUTTON_SIZE_Y * 1.5F), Formatting.GREEN);
         }