+++ src/main/java/com/robertx22/age_of_exile/mixins/OnCraftMixin.java
@@ -1,6 +1,7 @@
 package com.robertx22.age_of_exile.mixins;
 
 import com.robertx22.age_of_exile.mixin_methods.OnItemCrafted;
+import com.robertx22.age_of_exile.player_skills.OnCraftGiveSkillExp;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
@@ -17,6 +18,7 @@ public class OnCraftMixin {
         try {
             ItemStack stack = (ItemStack) (Object) this;
             OnItemCrafted.onCraft(stack, world, player, amount, ci);
+            OnCraftGiveSkillExp.onCraft(stack, world, player, amount, ci);
         } catch (Exception e) {
             e.printStackTrace();
         }