+++ src/main/java/com/robertx22/age_of_exile/saveclasses/player_skills/PlayerSkillEnum.java
@@ -7,6 +7,7 @@ public enum PlayerSkillEnum {
 
     MINING("mining", Words.Mining, Formatting.GRAY),
     FARMING("farming", Words.Farming, Formatting.YELLOW),
+    ALCHEMY("alchemy", Words.Alchemy, Formatting.LIGHT_PURPLE),
     NONE("none", Words.Elemental_Attack_Damage, Formatting.BLACK);
 
     public String id;